+++ DAQ_System/Source/Adapter/Interfaces/igyroscope.hpp
